requirements.txt
# ─────────────────────────────────────────
#  YouTube Summarizer — Requirements
#  Install: pip install -r requirements.txt
# ─────────────────────────────────────────

# UI
streamlit==1.35.0

# Backend API (optional — only needed for FastAPI server mode)
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0

# YouTube transcript extraction
youtube-transcript-api==0.6.2
yt-dlp==2024.5.27

#AI — Google Gemini (FREE: 1500 req/day, no credit card)
# Get key at: https://aistudio.google.com/app/apikey
google-generativeai==0.7.2

# Document Generation
reportlab==4.2.0
markdown2==2.4.13
Pillow==10.4.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
tiktoken==0.7.0